requirements.txt
wordcloud==1.9.4
matplotlib==3.10.0
arabic_reshaper==3.0.0
python-bidi==0.6.6
